Integral Tank Clean up
